Generate nodejs types using the nodejs lang generator.
