+++ pkg/codegen/docs/utils.go
@@ -22,6 +22,7 @@ import (
 	"strings"
 	"unicode"
 
+	"github.com/pulumi/pulumi/pkg/codegen/python"
 	"github.com/pulumi/pulumi/pkg/util/contract"
 )
 
@@ -55,4 +56,15 @@ func tokenToName(tok string) string {
 	components := strings.Split(tok, ":")
 	contract.Assertf(len(components) == 3, "malformed token %v", tok)
 	return title(components[2])
+}
+
+func getLanguagePropertyName(name string, lang string) string {
+	switch lang {
+	case "python":
+		return python.PyName(name)
+	case "go", "csharp":
+		return title(name)
+	default:
+		return wbr(name)
+	}
 }
